interview-to-bpmn/requirements.txt
# Audio processing
pydub>=0.25.1

# Transcription (local mode)
faster-whisper>=1.0.0
whisperx>=3.1.0
pyannote.audio>=3.1.0

# AI analysis
anthropic>=0.40.0
langchain>=0.3.0
langchain-anthropic>=0.3.0

# BPMN generation
lxml>=5.0.0

# Document generation
docxtpl>=0.18.0
python-docx>=1.1.0

# Web interface
streamlit>=1.40.0

# Utilities
pyyaml>=6.0
rich>=13.0.0
typer>=0.12.0

# OpenAI (for API transcription mode)
openai>=1.50.0
